refactor(tests): split review-police.test.ts under the file cap
At 1419 lines it was the largest test file in the repo and 419 past the
1000-line limit coding-police enforces on every write.

The 119 tests move to tests/review-police/ unchanged — same names, same
bodies, same assertions — split across eight files by their existing
describe groups, over four shared modules: constants, merge-command
fixtures, the temp repo and fake forge, the runHook probe, and the
strict-policy record builders. Every test file imports the wrapped test()
rather than bun's, so a failing assertion still carries the last hook
call's diagnostics; the three self-proof cases assert that for the new
layout.

Five lines outside the test bodies changed. Two resolve paths that used
import.meta.dir from the old file's directory. Three set forge fixture
state through setters, because an ES module cannot assign to an imported
binding.

moon.yml, scripts/check-test-slices.ts, tests/test-slices.test.ts,
tests/review-gate.test.ts and the review-governance zone in
.agentkit/review-policy.json follow the files, so the split stays in the
same CI slice and the same critical risk tier.

Refs #451
