+++ moon.yml
@@ -57,7 +57,7 @@ fileGroups:
     - "tests/install-tools.test.ts"
     - "tests/review-disciplines.test.ts"
     - "tests/review-gate.test.ts"
-    - "tests/review-police.test.ts"
+    - "tests/review-police/**/*"
     - "tests/review-profile.test.ts"
     - "tests/test-slices.test.ts"
     - "tools/review-gate"
@@ -224,7 +224,7 @@ fileGroups:
     - "tests/product-command.test.ts"
     - "tests/review-disciplines.test.ts"
     - "tests/review-gate.test.ts"
-    - "tests/review-police.test.ts"
+    - "tests/review-police/**/*"
     - "tests/review-profile.test.ts"
     - "tools/review-gate"
     - "tools/review-profile"
