+++ scripts/check-test-slices.ts
@@ -103,7 +103,14 @@ export const TEST_SLICES = {
     'tests/product-command.test.ts',
     'tests/review-disciplines.test.ts',
     'tests/review-gate.test.ts',
-    'tests/review-police.test.ts',
+    'tests/review-police/bypasses.test.ts',
+    'tests/review-police/fail-closed.test.ts',
+    'tests/review-police/forge-binding.test.ts',
+    'tests/review-police/police-hooks-parse.test.ts',
+    'tests/review-police/probe-diagnostics.test.ts',
+    'tests/review-police/probe-table.test.ts',
+    'tests/review-police/semantics.test.ts',
+    'tests/review-police/target-policy.test.ts',
     'tests/review-profile.test.ts',
   ],
   session: ['tests/session/agent-session.test.ts', 'tests/session/install-session-slice.test.ts'],
