+++ tests/review-gate.test.ts
@@ -292,7 +292,8 @@ describe('review-gate strict evidence validation', () => {
       'instructions/coding-discipline.md',
       'docs/review-process.md',
       'tests/review-gate.test.ts',
-      'tests/review-police.test.ts',
+      'tests/review-police/semantics.test.ts',
+      'tests/review-police/fixture.ts',
       'tests/review-disciplines.test.ts',
       'tests/agentkit-plugin.test.ts',
       'tests/codex-review-hooks.test.ts',
